More cocoaTouch

+++ sys/Graphics_colour.cpp
@@ -105,7 +105,7 @@ void _Graphics_setColour (Graphics graphics, Graphics_Colour colour) {
 			DeleteObject (my d_winBrush);
 			my d_winBrush = CreateSolidBrush (my d_winForegroundColour);
         #elif cocoaTouch
-            CGFloat components[] = {colour.red, colour.green, colour.blue, 1.0};
+            CGFloat components[] = {(float)colour.red, (float)colour.green, (float)colour.blue, 1.0};
             my d_macColour = CGColorCreate(CGColorSpaceCreateDeviceRGB(), components);
 		#elif mac
             my d_macColour. red = colour. red * 65535;
@@ -142,7 +142,7 @@ void _Graphics_setGrey (Graphics graphics, double fgrey) {
 			my d_winBrush = CreateSolidBrush (my d_winForegroundColour);
         #elif cocoaTouch
             if (fgrey < 0.0) fgrey = 0.0; else if (fgrey > 1.0) fgrey = 1.0;
-            CGFloat components[] = {fgrey, fgrey, fgrey};
+            CGFloat components[] = {(float)fgrey, (float)fgrey, (float)fgrey};
             my d_macColour = CGColorCreate(CGColorSpaceCreateDeviceRGB(), components);
 
 		#elif mac
@@ -271,6 +271,8 @@ static void highlight2 (Graphics graphics, long x1DC, long x2DC, long y1DC, long
 				CGContextSaveGState (my d_macGraphicsContext);
 
 				NSCAssert (my d_macGraphicsContext, @"nil context");
+				//CGContextTranslateCTM (my d_macGraphicsContext, 0, drawingArea. bounds. size. height);
+				//CGContextScaleCTM (my d_macGraphicsContext, 1.0, -1.0);
 
 				NSRect upperRect = NSMakeRect (x1DC, y2DC, x2DC - x1DC, y2DC_inner - y2DC);
 				NSRect leftRect  = NSMakeRect (x1DC, y2DC_inner, x1DC_inner - x1DC, y1DC_inner - y2DC_inner);
